fix CI
