+++ Makefile
@@ -184,11 +184,11 @@ build-core-test:
 	dune build ./_build/default/src/tests/test.exe
 
 #coupling: this is run by .github/workflow/tests.yml
-.PHONY: core-e2etest
-core-e2etest:
+.PHONY: core-test-e2e
+core-test-e2e:
 	SEMGREP_CORE=$(PWD)/bin/semgrep-core \
 	$(MAKE) -C interfaces/semgrep_interfaces test
-	python3 tests/e2e/test_target_file.py
+	python3 tests/semgrep-core-e2e/test_target_file.py
 
 ###############################################################################
 # External dependencies installation targets
